fixes for game version 409
This is a quick fix for augments which broke, the solution is not elegant but it works for now.

Moved the webhook options into usersettings.py, as well as added safeguards in case they are not set to avoid errors for everyone not using it.

Will add a check for unset options in the near future so you will get notified when new options are added and they are missing from your local options. 

Also added TOP_LEFT_COLOR to ngucon, as this was changed with the new update as well.