Improved class logic
Scripts are now designed as one-shots and stat tracking is done outside the script because it makes more sense to only do that in one place, instead of in each script. An added benefit of the one-shot alteration is that the script doesn't have to be restarted in order to change settings (besides inventory merge and boost).

Improved the inventory selector to use a button for setup and it's now stopping the user from editing the selected slots while a script is running, because it has to navigate to the inventory to grab the bitmap, which will interfere with the running script.